requirements.txt
beautifulsoup4
requests
sentence-transformers
faiss-cpu
groq
streamlit
python-dotenv
pdfminer.six
lxml
ddgs
